requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.5.0
anyio==3.6.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.11.2
blinker==1.6.3
cachetools==5.3.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.0.1
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.8
configparser==6.0.0
cssselect==1.2.0
cssutils==2.9.0
diff-match-patch==20230430
distro==1.8.0
dnspython==2.3.0
email-validator==2.1.0.post1
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.1.3
fastapi==0.92.0
filelock==3.13.1
Flask==3.0.0
fpdf==1.7.2
frozenlist==1.4.0
future==0.18.3
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==0.18.0
httplib2==0.22.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
importlib-metadata==6.8.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
Levenshtein==0.20.9
looseversion==1.3.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
networkx==3.2.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.26.2
openai==0.28.1
openpyxl==3.1.2
orjson==3.9.10
outcome==1.3.0
packaging==23.2
pandas==2.1.3
pathlib==1.0.1
Pillow==10.1.0
premailer==3.10.0
protobuf==4.25.1
prov==2.0.0
psutil==5.9.6
pyarrow==14.0.1
pycparser==2.21
pydantic==1.10.5
pydantic_core==2.6.3
pydeck==0.8.1b0
pydot==1.4.2
Pygments==2.16.1
pymongo==4.3.3
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pyparsing==3.1.1
PyQt6==6.6.0
PyQt6-Qt6==6.6.0
PyQt6-sip==13.6.0
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==1.0.1
python-dotenv==1.0.0
python-Levenshtein==0.20.9
python-multipart==0.0.6
pytz==2023.3.post1
pywin32==306
pyxnat==1.6
PyYAML==6.0.1
rapidfuzz==2.13.7
rdflib==7.0.0
redlines==0.4.2
referencing==0.31.0
requests==2.28.2
rich==13.6.0
rich-click==1.6.1
rpds-py==0.13.1
scipy==1.11.4
selenium==4.14.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
starlette==0.25.0
streamlit==1.28.2
tabula-py==2.9.0
tenacity==8.2.3
thefuzz==0.19.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traits==6.3.2
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.2
ujson==5.8.0
urllib3==1.26.14
uvicorn==0.23.2
validators==0.22.0
watchdog==3.0.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==3.0.1
wsproto==1.2.0
yagmail==0.15.293
yarl==1.9.2
zipp==3.17.0

opencv-python~=4.8.1.78